Refactor versioning apis and UIs
