+++ src/lib/layouts/workflow-run-layout.svelte
@@ -10,11 +10,7 @@
   import WorkflowHeader from '$lib/layouts/workflow-header.svelte';
   import { toDecodedPendingActivities } from '$lib/models/pending-activities';
   import { fetchAllEvents } from '$lib/services/events-service';
-  import {
-    getPollers,
-    getTaskQueueCompatibility,
-    getTaskQueueRules,
-  } from '$lib/services/pollers-service';
+  import { getPollers } from '$lib/services/pollers-service';
   import { fetchWorkflow } from '$lib/services/workflow-service';
   import { authUser } from '$lib/stores/auth-user';
   import { eventFilterSort, type EventSortOrder } from '$lib/stores/event-view';
@@ -26,32 +22,11 @@
     workflowRun,
   } from '$lib/stores/workflow-run';
   import type { NetworkError } from '$lib/types/global';
-  import type { WorkflowExecution } from '$lib/types/workflows';
 
   $: ({ namespace, workflow: workflowId, run: runId } = $page.params);
 
   let workflowError: NetworkError;
   let eventHistoryController: AbortController;
-
-  const getCompatibility = async (
-    workflow: WorkflowExecution,
-    taskQueue: string,
-  ) => {
-    const workflowUsesVersioning =
-      workflow?.mostRecentWorkerVersionStamp?.useVersioning;
-    if (!workflowUsesVersioning) return;
-    return await getTaskQueueCompatibility({
-      queue: taskQueue,
-      namespace,
-    });
-  };
-
-  const getRules = async (workflow: WorkflowExecution, taskQueue: string) => {
-    const workflowUsesVersioning =
-      workflow?.mostRecentWorkerVersionStamp?.useVersioning;
-    if (!workflowUsesVersioning) return;
-    return await getTaskQueueRules({ namespace, queue: taskQueue });
-  };
 
   const getWorkflowAndEventHistory = async (
     namespace: string,
@@ -74,8 +49,6 @@
 
     const { taskQueue } = workflow;
     const workers = await getPollers({ queue: taskQueue, namespace });
-    const compatibility = await getCompatibility(workflow, taskQueue);
-    const rules = await getRules(workflow, taskQueue);
 
     workflow.pendingActivities = await toDecodedPendingActivities(
       workflow,
@@ -83,8 +56,7 @@
       settings,
       $authUser?.accessToken,
     );
-    $workflowRun = { workflow, workers, rules, compatibility };
-
+    $workflowRun = { workflow, workers };
     eventHistoryController = new AbortController();
     $fullEventHistory = await fetchAllEvents({
       namespace,
