+++ src/lib/stores/workflow-run.ts
@@ -1,10 +1,6 @@
 import { writable } from 'svelte/store';
 
-import type {
-  GetPollersResponse,
-  TaskQueueCompatibility,
-  TaskQueueRules,
-} from '$lib/services/pollers-service';
+import type { GetPollersResponse } from '$lib/services/pollers-service';
 import { persistStore } from '$lib/stores/persist-store';
 import type { WorkflowExecution } from '$lib/types/workflows';
 
@@ -13,8 +9,6 @@ export const refresh = writable(0);
 export type WorkflowRunWithWorkers = {
   workflow: WorkflowExecution | null;
   workers: GetPollersResponse;
-  rules?: TaskQueueRules;
-  compatibility?: TaskQueueCompatibility;
 };
 
 export const initialWorkflowRun: WorkflowRunWithWorkers = {
