+++ src/routes/(app)/namespaces/[namespace]/task-queues/[queue]/+page.svelte
@@ -2,10 +2,10 @@
   import { page } from '$app/stores';
 
   import PageTitle from '$lib/components/page-title.svelte';
-  import TaskQueueWorkers from '$lib/pages/task-queue-workers.svelte';
+  import TaskQueue from '$lib/pages/task-queue.svelte';
 
   const queue = $page.params.queue;
 </script>
 
 <PageTitle title={`Task Queue | ${queue}`} url={$page.url.href} />
-<TaskQueueWorkers />
+<TaskQueue />
