+++ src/routes/(app)/namespaces/[namespace]/workflows/[workflow]/[run]/workers/+page.svelte
@@ -4,16 +4,18 @@
   import PageTitle from '$lib/components/page-title.svelte';
   import { translate } from '$lib/i18n/translate';
   import WorkflowPaddedLayout from '$lib/layouts/workflow-padded-layout.svelte';
-  import WorkflowWorkers from '$lib/pages/workflow-workers.svelte';
+  import TaskQueueWorkers from '$lib/pages/task-queue-workers.svelte';
+  import { workflowRun } from '$lib/stores/workflow-run';
 
-  const workflow = $page.params.workflow;
+  $: ({ workflow: workflowId } = $page.params);
+  $: ({ workers, workflow } = $workflowRun);
 </script>
 
 <PageTitle
-  title={`${translate('workflows.workers-tab')} | ${workflow}`}
+  title={`${translate('workflows.workers-tab')} | ${workflowId}`}
   url={$page.url.href}
 />
 
 <WorkflowPaddedLayout>
-  <WorkflowWorkers />
+  <TaskQueueWorkers {workers} taskQueue={workflow?.taskQueue} />
 </WorkflowPaddedLayout>
